redirect customers

+++ cse135/src/DeleteProduct.java
@@ -27,11 +27,15 @@ public class DeleteProduct extends HttpServlet {
 	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		int id = Integer.parseInt(request.getParameter("id"));
-		ProductDao dao = new ProductDao();
-		dao.deleteProduct(id);
-		response.getWriter().print("Product deleted");
-		System.out.println("deleting");
+		if (Browse.authOwner(request, response)) {
+			response.sendRedirect("browse");
+		} else {
+			int id = Integer.parseInt(request.getParameter("id"));
+			ProductDao dao = new ProductDao();
+			dao.deleteProduct(id);
+			response.getWriter().print("Product deleted");
+			System.out.println("deleting");
+		}
 	}
 
 	/**
